Add dotfiles testing github workflow
This uncovered a few things that got fixed to get the workflow passing:
- the boolean expressions within the dotfiles script were not working
  correctly. Fixed by checking for explicit boolean values
- sudo was mising and is required in a few places
- sourcing '.bash_profile' in the dotfiles script after linking dotfiles
  would fail due to set -e, because some of the 'bind' funtions in
bash_profile throw warnings when run called in a non-interactive shell.
The simple, and probably better, fix is to source .bashrc instead.
However, '|| true' had to be added to bash rc to prevent it from causing
dotfiles to fail with set -e, since it would have a error return status
otherwise
- the other changes are just to avoid a few paths when being run from
  the CI workflow